Refactoring Currency.Service

+++ Currency.Service/Currency.Service.API/Entities/Currency.cs
@@ -3,7 +3,7 @@ namespace Currency.Service.API.Entities
     public class Currency
     {
         public string CurrencyCode { get; set; }
-        public long CurrencyValue { get; set; }
-        public long Scale { get; set; }
+        public int CurrencyValue { get; set; }
+        public int Scale { get; set; }
     }
 }
